Fix warning

+++ src/bmpsavefile.c
@@ -45,7 +45,7 @@ static void calculateHeader(BMP *bmp, BMP_FORMAT format, uint32_t *offset, uint3
     bmp->header.compression = 0;
     bmp->header.clrUsed = 0;
     bmp->header.clrImportant = 0;
-    bmp->header.imageSize = rowSize * height;
+    bmp->header.imageSize = ((uint32_t) rowSize) * ((uint32_t) height);
 
     *offset = (uint32_t)(14L + bmp->header.headerSize);
     *fileSize = (uint32_t)(*offset + bmp->header.imageSize);
